+++ src/lib.rs
@@ -188,7 +188,6 @@
 //! - `defmt-0.3` - unstable feature which adds [`defmt::Format`] impl for [`Error`].
 
 #![no_std]
-#![doc(html_root_url = "https://docs.rs/nb/1.0.0")]
 
 use core::fmt;